+++ azure_data_cosmos_engine/src/query/producer/state.rs
@@ -84,7 +84,9 @@ impl PartitionState {
     /// Gets the next [`DataRequest`] for this partition, if one is needed.
     pub fn request(&self) -> Option<DataRequest> {
         match &self.stage {
-            PaginationState::Initial => Some(DataRequest::new(0, self.pkrange.id.clone(), None)),
+            PaginationState::Initial => {
+                Some(DataRequest::new(0, self.pkrange.id.clone(), None, None))
+            }
             PaginationState::Continuing {
                 next_page_index,
                 token,
@@ -92,6 +94,65 @@ impl PartitionState {
                 *next_page_index as u64,
                 self.pkrange.id.clone(),
                 Some(token.clone()),
+                None,
+            )),
+            PaginationState::Done => None,
+        }
+    }
+
+    pub fn update_state(&mut self, continuation: Option<String>) {
+        self.stage.update(continuation);
+    }
+
+    pub fn started(&self) -> bool {
+        !matches!(self.stage, PaginationState::Initial)
+    }
+
+    pub fn done(&self) -> bool {
+        matches!(self.stage, PaginationState::Done)
+    }
+}
+
+#[derive(Debug)]
+pub struct QueryChunkState {
+    /// The index of the query chunks in the list used by the pipeline.
+    pub index: usize,
+    /// The partition key range id this state is for.
+    pub pkrange_id: String,
+    /// The query that belongs to this set of item identities.
+    pub query: String,
+    /// The current stage of pagination for this partition.
+    pub stage: PaginationState,
+}
+
+impl QueryChunkState {
+    /// Initializes a partition state for the given partition key range.
+    pub fn new(index: usize, pkrange_id: String, query: String) -> Self {
+        Self {
+            index: index,
+            pkrange_id: pkrange_id,
+            query: query,
+            stage: PaginationState::Initial,
+        }
+    }
+
+    /// Gets the next [`DataRequest`] for this partition, if one is needed.
+    pub fn request(&self) -> Option<DataRequest> {
+        match &self.stage {
+            PaginationState::Initial => Some(DataRequest::new(
+                self.index as u64,
+                self.pkrange_id.clone(),
+                None,
+                Some(self.query.clone()),
+            )),
+            PaginationState::Continuing {
+                next_page_index: _,
+                token,
+            } => Some(DataRequest::new(
+                self.index as u64,
+                self.pkrange_id.clone(),
+                Some(token.clone()),
+                Some(self.query.clone()),
             )),
             PaginationState::Done => None,
         }