+++ cosmoscx/src/pipeline.rs
@@ -3,9 +3,11 @@
 
 //! Functions related to creating and executing query pipelines.
 
+use std::str::FromStr;
+
 use azure_data_cosmos_engine::{
-    query::{PartitionKeyRange, QueryPipeline, QueryPlan},
-    ErrorKind,
+    query::{ItemIdentity, PartitionKeyRange, QueryPipeline, QueryPlan},
+    ErrorKind, PartitionKeyKind,
 };
 use serde::Deserialize;
 
@@ -44,6 +46,7 @@ impl Pipeline {
 /// Creates a new query pipeline from a JSON query plan and list of partitions.
 ///
 /// # Parameters
+/// - `query`: A [`Str`] containing the query to be executed.
 /// - `query_plan_json`: A [`Str`] containing the serialized query plan, as recieved from the gateway, in JSON.
 /// - `pkranges`: A [`Str`] containing the serialized partition key ranges list, as recieved from the gateway, in JSON.
 #[no_mangle]
@@ -80,6 +83,62 @@ pub extern "C" fn cosmoscx_v0_query_pipeline_create<'a>(
     }
 
     inner(query, query_plan_json, pkranges).into()
+}
+
+/// Creates the relevant partition-scoped queries for executing the read many operation along with the pipeline to run them.
+///
+/// # Parameters
+/// - `item_identities`: A [`Str`] containing the serialized item identities in JSON.
+/// - `pkranges`: A [`Str`] containing the serialized partition key ranges list, as received from the gateway, in JSON.
+/// - `pk_kind`: A [`Str`] containing the partition key kind.
+/// - `pk_version`: The partition key version.
+#[no_mangle]
+pub extern "C" fn cosmoscx_v0_readmany_pipeline_create<'a>(
+    item_identities: Str<'a>,
+    pkranges: Str<'a>,
+    pk_kind: Str<'a>,
+    pk_version: u8,
+    pk_paths: Str<'a>,
+) -> FfiResult<Pipeline> {
+    #[derive(Deserialize)]
+    struct PartitionKeyRangeResult {
+        #[serde(rename = "PartitionKeyRanges")]
+        pub ranges: Vec<PartitionKeyRange>,
+    }
+
+    fn inner<'a>(
+        item_identities: Str<'a>,
+        pkranges: Str<'a>,
+        pk_kind: Str<'a>,
+        pk_version: u8,
+        pk_paths: Str<'a>,
+    ) -> Result<Box<QueryPipeline>, azure_data_cosmos_engine::Error> {
+        let item_identities_json = unsafe { item_identities.as_str().not_null() }?;
+        let pkranges_json = unsafe { pkranges.as_str().not_null() }?;
+        let pk_kind_json = unsafe { pk_kind.as_str().not_null() }?;
+        let pk_paths_json = unsafe { pk_paths.as_str().not_null() }?;
+
+        let pkranges: PartitionKeyRangeResult = serde_json::from_str(pkranges_json)
+            .map_err(|e: serde_json::Error| ErrorKind::InvalidGatewayResponse.with_source(e))?;
+        let item_identities: Vec<ItemIdentity> = serde_json::from_str(item_identities_json)
+            .map_err(|e: serde_json::Error| ErrorKind::InvalidGatewayResponse.with_source(e))?;
+        // Deserialize pk_paths as a JSON array of strings from Go
+        let pk_paths_vec: Vec<String> = serde_json::from_str(pk_paths_json)
+            .map_err(|e: serde_json::Error| ErrorKind::InvalidGatewayResponse.with_source(e))?;
+
+        // SAFETY: We should no longer need either of the parameter slices, we copied them into owned data.
+        tracing::debug!(item_identities = ?item_identities, pkranges = ?pkranges.ranges, pk_kind = ?pk_kind_json, pk_version = ?pk_version, pk_paths = ?pk_paths_vec, "creating readmany pipeline");
+        let pipeline = QueryPipeline::for_read_many(
+            item_identities,
+            pkranges.ranges,
+            PartitionKeyKind::from_str(pk_kind_json).expect("invalid partition key kind"),
+            pk_version,
+            pk_paths_vec,
+        )?;
+        Ok(Box::new(pipeline))
+    }
+
+    inner(item_identities, pkranges, pk_kind, pk_version, pk_paths).into()
 }
 
 /// Frees the memory associated with a pipeline.
@@ -273,6 +332,5 @@ pub extern "C" fn cosmoscx_v0_query_pipeline_provide_data<'a>(
         }
         Ok(())
     }
-
     inner(pipeline, responses).into()
 }