+++ go/azcosmoscx/engine.go
@@ -113,3 +113,13 @@ func (p *clientEngineQueryPipeline) Run() (*queryengine.PipelineResult, error) {
 func (p *clientEngineQueryPipeline) ProvideData(results []queryengine.QueryResult) error {
 	return p.pipeline.ProvideData(results)
 }
+
+// CreateReadManyPipeline creates the relevant partition-scoped queries for executing the read many operation along with the pipeline to run them.
+func (e *nativeQueryEngine) CreateReadManyPipeline(items []queryengine.ItemIdentity, pkranges string, pkKind string, pkVersion int8, pkPaths []string) (queryengine.QueryPipeline, error) {
+	pipeline, err := newReadManyPipeline(items, pkranges, pkKind, pkVersion, pkPaths)
+	if err != nil {
+		return nil, err
+	}
+
+	return &clientEngineQueryPipeline{pipeline, "", false}, nil
+}